Moved EventLog files to separate project
