+++ Src/CrispyWaffle/CrispyWaffle.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
 		<ProjectGuid>9A19103F-16F7-4668-BE54-9A1E7A4F7556</ProjectGuid>
@@ -10,7 +10,6 @@
 
 	<ItemGroup>
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-		<PackageReference Include="System.Diagnostics.EventLog" Version="[6.0.0,7.0.0)" />
 		<PackageReference Include="System.Security.Permissions" Version="[6.0.0,7.0.0)" />
 		<PackageReference Include="System.Text.Json" Version="8.0.3" />
 	</ItemGroup>
