+++ Src/CrispyWaffle/Log/Adapters/ILogAdapter.cs
