--- Src/CrispyWaffle/Log/Providers/IEventIdProvider.cs
@@ -1,15 +1,0 @@
-﻿namespace CrispyWaffle.Log.Providers
-{
-    /// <summary>
-    /// Interface IEventIdProvider
-    /// </summary>
-    public interface IEventIdProvider
-    {
-        /// <summary>
-        /// Computes the event identifier.
-        /// </summary>
-        /// <param name="message">The message.</param>
-        /// <returns>System.UInt16.</returns>
-        ushort ComputeEventId(string message);
-    }
-}
