+++ Src/CrispyWaffle/Log/Providers/ILogProvider.cs
